fixed user restriction functions
