+++ routes/html-routes.js
@@ -1,4 +1,8 @@
-module.exports = function (app) {
+module.exports = function (
+  app,
+  isAuthenticatedMiddleware,
+  isNotAuthenticatedMiddleware
+) {
   app.get("/", isAuthenticatedMiddleware(), (req, res) => {
     console.log("///////CurrentUser: ", req.user);
     if (req.user === "admin") {
@@ -13,25 +17,3 @@ module.exports = function (app) {
     res.render("login");
   });
 };
-
-function isAuthenticatedMiddleware() {
-  return (req, res, next) => {
-    console.log(
-      `req.session.passport.user: ${JSON.stringify(req.session.passport)}`
-    );
-
-    if (req.isAuthenticated()) return next();
-    res.redirect("/login");
-  };
-}
-
-function isNotAuthenticatedMiddleware() {
-  return (req, res, next) => {
-    console.log(
-      `req.session.passport.user: ${JSON.stringify(req.session.passport)}`
-    );
-
-    if (!req.isAuthenticated()) return next();
-    res.redirect("/");
-  };
-}
